refactor: extract CodeBlock subsystem to models/code_block.py
Pure move of the CodeBlock sandbox/secure-runner surface (Credential,
CodeBlockStep, CodeBlockOTPError, CodeBlock, and the OTP-binding helpers)
out of block.py into a sibling module. block.py stays the import facade and
re-exports every name via a late import, so there are zero call-site changes.

Trimmed the now-unused imports left behind in block.py.
